Fix Catch2
Yeah, header only boost was a problem

+++ include/Utilities.hpp
@@ -401,7 +401,7 @@ template <typename FloatingPointType>
       { return static_cast<int_fast64_t>(2 * simplices_per_timeslice); }
       else if (simplices < 1000)
       {
-        return static_cast<int_fast64_t>(0.4 * simplices_per_timeslice);
+        return 0.4 * simplices_per_timeslice;
       }
       else if (simplices < 10000)
       {
